Move Action menu to own file
