+++ frontend/src/DeviceList.js
@@ -35,6 +35,7 @@ import {
 import PropTypes from "prop-types";
 import GenerateData from "./SampleData";
 import generateHeadCells from "./headCells";
+import ActionMenu from "./ActionMenu";
 // imports for material ui & etc
 
 // temporary row
@@ -166,36 +167,7 @@ function rowExtras(extras) {
     }
     return extraStr;
 }
-function ActionMenu() {
-    // will need updating to function with individual devices
-    const [anchorEl, setAnchorEl] = React.useState(null);
 
-    const handleClick = (event) => {
-        setAnchorEl(event.currentTarget);
-    };
-
-    const handleClose = () => {
-        setAnchorEl(null);
-    };
-    return (
-        <React.Fragment>
-            <IconButton aria-controls="simple-menu" aria-haspopup="true" onClick={handleClick}>
-                <MoreVert />
-            </IconButton>
-            <Menu
-                id="simple-menu"
-                anchorEl={anchorEl}
-                keepMounted
-                open={Boolean(anchorEl)}
-                onClose={handleClose}
-            >
-                <MenuItem onClick={handleClose}>View details</MenuItem>
-                <MenuItem onClick={handleClose}>Start stream with this as receiver</MenuItem>
-                <MenuItem onClick={handleClose}><span class="warningText">Delete</span></MenuItem>
-            </Menu>
-        </React.Fragment>
-    );
-}
 
 function DevicesTable() {
     var data = rows[arguments[0]];
